requirements.txt
django
djangorestframework
black